# back/internal/api/handlers/auth.go
package handlers

import (
	"database/sql"
	"net/http"

	"back/internal/config"
	"back/internal/database/models"
	"back/internal/services"
	"back/internal/utils"

	"github.com/gin-gonic/gin"
	"github.com/go-playground/validator/v10"
)

type AuthHandler struct {
	db          *sql.DB
	config      *config.Config
	validator   *validator.Validate
	authService *services.AuthService
}

func NewAuthHandler(db *sql.DB, cfg *config.Config) *AuthHandler {
	return &AuthHandler{
		db:          db,
		config:      cfg,
		validator:   validator.New(),
		authService: services.NewAuthService(db, cfg),
	}
}

// Signup maneja el registro de nuevos usuarios
// @Summary Registrar nuevo usuario
// @Description Registra un nuevo usuario en la plataforma ANB Rising Stars
// @Tags auth
// @Accept json
// @Produce json
// @Param user body models.UserRegistration true "Datos del usuario a registrar"
// @Success 201 {object} models.APIResponse
// @Failure 400 {object} models.APIResponse
// @Failure 500 {object} models.APIResponse
// @Router /auth/signup [post]
func (h *AuthHandler) Signup(c *gin.Context) {
	var req models.UserRegistration

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, models.APIResponse{
			Error: "Invalid request format",
		})
		return
	}

	// Validar datos de entrada
	if err := h.validator.Struct(req); err != nil {
		c.JSON(http.StatusBadRequest, models.APIResponse{
			Error: "Validation failed: " + err.Error(),
		})
		return
	}

	// Validar que las contraseñas coincidan
	if req.Password1 != req.Password2 {
		c.JSON(http.StatusBadRequest, models.APIResponse{
			Error: "Passwords do not match",
		})
		return
	}

	// Verificar si el email ya existe
	if exists, err := h.authService.EmailExists(req.Email); err != nil {
		c.JSON(http.StatusInternalServerError, models.APIResponse{
			Error: "Internal server error",
		})
		return
	} else if exists {
		c.JSON(http.StatusBadRequest, models.APIResponse{
			Error: "Email already registered",
		})
		return
	}

	// Hash de la contraseña
	hashedPassword, err := utils.HashPassword(req.Password1)
	if err != nil {
		c.JSON(http.StatusInternalServerError, models.APIResponse{
			Error: "Internal server error",
		})
		return
	}

	// Crear usuario
	user := &models.User{
		FirstName:    req.FirstName,
		LastName:     req.LastName,
		Email:        req.Email,
		PasswordHash: hashedPassword,
		City:         req.City,
		Country:      req.Country,
		//CreatedAt:    time.Now(), fecha tomada por BD
		//UpdatedAt:    time.Now(),
	}

	if err := h.authService.CreateUser(user); err != nil {
		c.JSON(http.StatusInternalServerError, models.APIResponse{
			Error: "Failed to create user",
		})
		return
	}

	c.JSON(http.StatusCreated, models.APIResponse{
		Message: "User created successfully",
	})
}

// Login maneja la autenticación de usuarios
// @Summary Autenticar usuario
// @Description Autentica un usuario existente y devuelve un token JWT
// @Tags auth
// @Accept json
// @Produce json
// @Param credentials body models.UserLogin true "Credenciales del usuario"
// @Success 200 {object} models.LoginResponse
// @Failure 400 {object} models.APIResponse
// @Failure 401 {object} models.APIResponse
// @Failure 500 {object} models.APIResponse
// @Router /auth/login [post]
func (h *AuthHandler) Login(c *gin.Context) {
	var req models.UserLogin

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, models.APIResponse{
			Error: "Invalid request format",
		})
		return
	}

	// Validar datos de entrada
	if err := h.validator.Struct(req); err != nil {
		c.JSON(http.StatusBadRequest, models.APIResponse{
			Error: "Validation failed: " + err.Error(),
		})
		return
	}

	// Buscar usuario por email
	user, err := h.authService.GetUserByEmail(req.Email)
	if err != nil {
		c.JSON(http.StatusUnauthorized, models.APIResponse{
			Error: "Invalid credentials",
		})
		return
	}

	// Verificar contraseña
	if !utils.CheckPassword(req.Password, user.PasswordHash) {
		c.JSON(http.StatusUnauthorized, models.APIResponse{
			Error: "Invalid credentials",
		})
		return
	}

	// Generar JWT token
	token, err := utils.GenerateJWT(user.ID, user.Email, h.config.JWTSecret, h.config.JWTExpiration)
	if err != nil {
		c.JSON(http.StatusInternalServerError, models.APIResponse{
			Error: "Failed to generate token",
		})
		return
	}

	response := models.LoginResponse{
		AccessToken: token,
		TokenType:   "Bearer",
		ExpiresIn:   int(h.config.JWTExpiration.Seconds()),
	}

	c.JSON(http.StatusOK, response)
}

// GetProfile retorna el perfil del usuario autenticado
// @Summary Obtener perfil de usuario
// @Description Obtiene el perfil del usuario autenticado
// @Tags auth
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} models.User
// @Failure 401 {object} models.APIResponse
// @Failure 404 {object} models.APIResponse
// @Router /auth/profile [get]
func (h *AuthHandler) GetProfile(c *gin.Context) {
	userID, exists := c.Get("user_id")
	if !exists {
		c.JSON(http.StatusUnauthorized, models.APIResponse{
			Error: "User not authenticated",
		})
		return
	}

	user, err := h.authService.GetUserByID(int(userID.(int64)))
	if err != nil {
		c.JSON(http.StatusNotFound, models.APIResponse{
			Error: "User not found",
		})
		return
	}

	c.JSON(http.StatusOK, user)
}

// Logout maneja el cierre de sesión
// @Summary Cerrar sesión
// @Description Cierra la sesión del usuario autenticado
// @Tags auth
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} models.APIResponse
// @Failure 401 {object} models.APIResponse
// @Router /auth/logout [post]
func (h *AuthHandler) Logout(c *gin.Context) {

	c.JSON(http.StatusOK, models.APIResponse{
		Message: "Logged out successfully",
	})
}